move fixed_assets out from accounting namespace
