+++ app/views/fixed_assets/commodities/_create_form.html.erb
@@ -25,26 +25,12 @@
       <th><%= f.label :selite, t('Selite') %></th>
       <td><%= f.text_field :selite %></td>
     </tr>
-    <tr>
-      <th><%= f.label :summa, t('Hankintasumma') %></th>
-      <td><%= f.text_field :summa %></td>
-    </tr>
+
     <tr>
       <th><%= f.label :hankintapvm, t('Hankintapvm') %></th>
       <td><%= f.date_field :hankintapvm %></td>
     </tr>
-    <tr>
-      <th><%= f.label :kustp, t('Kustannuspaikka') %></th>
-      <td><%= f.text_field :kustp %></td>
-    </tr>
-    <tr>
-      <th><%= f.label :kohde, t('Kohde') %></th>
-      <td><%= f.text_field :kohde %></td>
-    </tr>
-    <tr>
-      <th><%= f.label :projekti, t('Projekti') %></th>
-      <td><%= f.text_field :projekti %></td>
-    </tr>
+
     <tr>
       <th><%= t('Laatija') %></th>
       <td><%= @commodity.laatija %></td>
