+++ app/views/fixed_assets/commodities/_form.html.erb
