+++ app/views/fixed_assets/commodities/new.html.erb
