+++ app/views/fixed_assets/commodities/select_purchase_order.html.erb
