+++ app/views/fixed_assets/commodities/select_voucher.html.erb
